MDL-40581 backup: Convert backupids cache.
Most calls in Moodle restore are to newitem and parentitemid.
The new cache stores those items to allow restore to use
the cache in most places without asking the database for answers.

We include an insert counter to allow us to know that something isn't
in the database if we haven't filled the cache yet.