MDL-40581 backup: Update unit test for cache
Unit tests were adjusted as records now come from the database
rather than directly from the cache.  So in many cases they will
not be integer as the db returns them as strings.

Due to the cache answering questions without contacting the database
the test has to be adjusted to return expected results because of that
behaviour.  get_backup_ids_mappings was used to test caches as
get_backup_ids_record is no longer cached at all.