MDL-53875 cache: Ensure session cache mode clearing works.
Sessions in PHP are limited to a-zA-Z0-9,- so the definition of
simplekeys has been expanded to include that.  This aligns the session
keyspace for prefixes including sessions with the ability for caches like
the filesystem to manage them.  The original restriction set was introduced
in MDL-36120 and was an arbritary decision of key selection.  Likely based
on a filesystem store to reduce the hasing requirements.

Session id's are used in full now which allows for better matching, prefixes
and session keys align to allow for a search of the keyspace to purge unused keys.

The clearing of historical session data can only be completed when the session has
been removed from the database sessions table.  Doing it inline would produce too
many difficulties with performance on large sites.  Large sites should run the cron
task more frequently.