MDL-43507 cache: Only set static_acceleration if required in set_many
Static acceleration can be of a fixed size.  If you happen to call set_many
and there are more elements than static_acceleration can store it cycles
through the values with a lot of overhead.  This removes that cycling and
only adds the values that will stay.

This was discovered while working on PHPUnit performance and using set_many
with large values for the purpose of databasemeta.