MDL-43507 testing: Use cache for events in PHPUnit.
Caches are reset during each run on PHPUnit and provide
performance boosts when in use. They are enabled to mimic reality
as well as provide a faster testing experience.  Files on disk do not
change during a testing run, and if a test alters them it needs to
make the appropriate adjustments.