MDL-56273 caching: Purge active caches when purging all.
Not all currently active caches are found by a purge_all()
request.  We also loop through currently active caches to ensure
they are purged as well.  Caches much be purged before
the cache factory is reset.  Otherwise we don't know what
caching was setup when we attempt to complete a full purge.