MDL-56624 plugins: Allow dependent plugins to be uninstalled
When a plugin that is required by this plugin is in the
NEW status, we should all the other plugins to be removed.
In a situation where an entire set of plugins is to be removed,
you want to be able to remove all of them in the right order and
then do a single code release.  Not all installs can delete the plugin
code one-by-one as you go.